@@ -1,12 +1,12 @@
-require "texticle/searchable"
+require "textacular/searchable"
 
 class Project < ActiveRecord::Base
   belongs_to :chapter
   belongs_to :hidden_by_user, class_name: "User"
   has_many :votes
   has_many :users, :through => :votes
-  has_many :photos, :order => "photos.sort_order asc, photos.id asc"
-  has_many :real_photos, :order => "photos.sort_order asc, photos.id asc", :conditions => proc { Photo.arel_table[:image_content_type].matches('image/%') }, :class_name => "Photo"
+  has_many :photos, -> { order(sort_order: :asc, id: :asc) }
+  has_many :real_photos, -> { where(Photo.arel_table[:image_content_type].matches('image/%')).order(sort_order: :asc, id: :asc) }, class_name: "Photo"
 
   attr_accessible :name, :title, :url, :email, :phone, :about_me, :about_project,
                   :chapter_id, :extra_question_1, :extra_question_2, :extra_question_3,
@@ -81,7 +81,7 @@ class Project < ActiveRecord::Base
 
   def self.recent_winners
     subquery = select("DISTINCT ON (chapter_id) projects.*").where("projects.funded_on IS NOT NULL").order(:chapter_id, :funded_on).reverse_order
-    select("*").from("(#{subquery.to_sql}) AS distinct_chapters").order(:funded_on).reverse_order
+    select("*").from("(#{subquery.to_sql}) AS distinct_chapters").order("distinct_chapters.funded_on").reverse_order
   end
 
   def self.csv_export(projects)
@@ -110,7 +110,7 @@ class Project < ActiveRecord::Base
   end
 
   def deliver_winning_email
-    mailer.winner(self).deliver
+    mailer.winner(self).deliver_now
   end
 
   def declare_winner!(new_chapter = nil)
