@@ -12,11 +12,11 @@ class User < ActiveRecord::Base
   validates_uniqueness_of :email
 
   has_many :roles
-  has_many :chapters, :through => :roles
-  has_many :dean_chapters, :source => :chapter, :through => :roles, :conditions => "roles.name = 'dean'"
+  has_many :chapters, through: :roles
+  has_many :dean_chapters, -> { where(roles: { name: 'dean' }) }, through: :roles, source: :chapter
 
   has_many :votes
-  has_many :projects, :through => :votes
+  has_many :projects, through: :votes
 
   def logged_in?
     true
