@@ -1,5 +1,6 @@
 class ApplicationController < ActionController::Base
-  include Clearance::Authentication
+  include Clearance::Controller
+
   protect_from_forgery
   before_filter :set_locale
   before_filter :fix_chapter_ids
