@@ -1,6 +1,6 @@
 class HomeController < ApplicationController
   def index
-    @chapters = Chapter.active.visitable.for_display.all
-    @projects = Project.recent_winners.limit(15).all
+    @chapters = Chapter.active.visitable.for_display
+    @projects = Project.recent_winners.limit(15)
   end
 end
