@@ -1,7 +1,7 @@
 class SubdomainConstraint
   def self.matches?(request)
     case request.subdomains.first
-    when 'www', '', nil, 'staging'
+    when 'www', '', nil, 'staging', 'preview'
       false
     else
       true
