@@ -14,7 +14,7 @@ HTTP_ERRORS = [Timeout::Error,
                Net::HTTPHeaderSyntaxError,
                Net::ProtocolError]
 
-SMTP_SERVER_ERRORS = [TimeoutError,
+SMTP_SERVER_ERRORS = [Timeout::Error,
                       IOError,
                       Net::SMTPUnknownError,
                       Net::SMTPServerBusy,
