@@ -1,15 +1,15 @@
 step 'I attach 3 photos' do
   photos = Rails.root.join("spec", "support", "fixtures")
-  find(".new_photo").set(photos + "1.jpg")
-  find(".new_photo").set(photos + "2.jpg")
-  find(".new_photo").set(photos + "3.jpg")
+  find(".new_photo").set(photos + "1.JPG")
+  find(".new_photo").set(photos + "2.JPG")
+  find(".new_photo").set(photos + "3.JPG")
   click_button("Save")
 end
 
 step 'I should see the three photos in the carousel' do
-  expect(page).to have_css("#project-gallery img[src*='1.jpg']")
-  expect(page).to have_css("#project-gallery img[src*='2.jpg']")
-  expect(page).to have_css("#project-gallery img[src*='3.jpg']")
+  expect(page).to have_css("#project-gallery img[src*='1.JPG']")
+  expect(page).to have_css("#project-gallery img[src*='2.JPG']")
+  expect(page).to have_css("#project-gallery img[src*='3.JPG']")
 end
 
 step 'I set the last image to be first' do
@@ -20,5 +20,5 @@ step 'I set the last image to be first' do
 end
 
 step 'I should see that last image when I load the page' do
-  expect(page).to have_css(".project-gallery img[src*='3.jpg']:not(.faded)")
+  expect(page).to have_css(".project-gallery img[src*='3.JPG']:not(.faded)")
 end
