@@ -1,6 +1,11 @@
 require 'spec_helper'
 
 describe 'layouts/application' do
+  before(:each) do
+    controller.request.path_parameters[:controller] = "home"
+    controller.request.path_parameters[:action] = "index"
+  end
+
   context 'the html tag' do
     it 'includes the default language locale' do
       view.stubs(:signed_in?).returns(false)
@@ -29,7 +34,7 @@ describe 'layouts/application' do
 
       # Can't figure out why have_selector isn't working here
       I18n.available_locales.each do |locale|
-        expect(rendered).to include("hreflang=\"#{locale}\" rel=\"alternate\"")
+        expect(rendered).to include("rel=\"alternate\" hreflang=\"#{locale}\"")
       end
     end
   end
